integrating  landing page with reservations
