+++ src/components/cards/ProfessorCard.js
@@ -1,48 +1,68 @@
 import { AccessTime, Business, MailOutline, Phone } from '@mui/icons-material';
 import { Card, CardActionArea, CardContent, CardMedia, Chip, List, ListItem, ListItemIcon, ListItemText, Typography } from '@mui/material';
+import Link from 'next/link';
+import PersonIcon from '@mui/icons-material/Person';
+import { getColor } from '@/utils/getColor';
 
 export default function ProfessorCard({ name, email, phone, office, officeHours, image, style, subjects }) {
+    const prepare_subjects = () => {
+        return subjects.map(subject => subject.name).join('-');
+    };
+
     return (
         <Card sx={{ maxWidth: 350, ...style }}>
-            <CardActionArea>
-                <CardMedia component='img' height='140' image='https://www.w3schools.com/howto/img_avatar.png' alt='Professor' />
-                <CardContent>
-                    <Typography gutterBottom variant='h5' component='div'>
-                        {name}
-                    </Typography>
-                    <List>
+            <Link
+                href={{
+                    pathname: '/reservations',
+                    query: { name: name, subject: prepare_subjects() },
+                }}
+                style={{ textDecoration: 'none' }}
+            >
+                <CardActionArea>
+                    <CardMedia component='img' height='140' image='https://www.w3schools.com/howto/img_avatar.png' alt='Professor' />
+                    <CardContent>
                         <ListItem>
                             <ListItemIcon>
-                                <MailOutline />
+                                <PersonIcon />
                             </ListItemIcon>
                             <Typography variant='body1' color='text.secondary'>
-                                {email}
+                                {name}
                             </Typography>
                         </ListItem>
-                        <ListItem>
-                            <ListItemIcon>
-                                <Phone />
-                            </ListItemIcon>
-                            <Typography variant='body1' color='text.secondary'>
-                                {phone}
-                            </Typography>
-                        </ListItem>
-                        <ListItem>
-                            <ListItemIcon>
-                                <Business />
-                            </ListItemIcon>
-                            <Typography variant='body1' color='text.secondary'>
-                                {office}
-                            </Typography>
-                        </ListItem>
-                        <ListItem sx={{ display: 'flex', gap: '0.5rem', flexWrap: 'wrap', justifyContent: 'center' }}>
-                            {subjects.map(subject => (
-                                <Chip key={subject.id} label={subject.name} />
-                            ))}
-                        </ListItem>
-                    </List>
-                </CardContent>
-            </CardActionArea>
+                        <List>
+                            <ListItem>
+                                <ListItemIcon>
+                                    <MailOutline />
+                                </ListItemIcon>
+                                <Typography variant='body1' color='text.secondary'>
+                                    {email}
+                                </Typography>
+                            </ListItem>
+                            <ListItem>
+                                <ListItemIcon>
+                                    <Phone />
+                                </ListItemIcon>
+                                <Typography variant='body1' color='text.secondary'>
+                                    {phone}
+                                </Typography>
+                            </ListItem>
+                            <ListItem>
+                                <ListItemIcon>
+                                    <Business />
+                                </ListItemIcon>
+                                <Typography variant='body1' color='text.secondary'>
+                                    {office}
+                                </Typography>
+                            </ListItem>
+                            <ListItem sx={{ display: 'flex', gap: '0.5rem', flexWrap: 'wrap', justifyContent: 'center' }}>
+                                {subjects.map(subject => (
+                                    <Chip key={subject.id} label={subject.name} sx={{ backgroundColor: getColor(subject.name) }} />
+                                ))}
+                            </ListItem>
+                        </List>
+                    </CardContent>
+                </CardActionArea>
+            </Link>
         </Card>
     );
 }
